+++ src/common/z-file.h
@@ -205,7 +205,7 @@ extern bool file_skip(ang_file *f, int bytes);
  * Reads n bytes from file 'f' into buffer 'buf'.
  * Returns Number of bytes read; -1 on error
  */
-extern int file_read(ang_file *f, char *buf, size_t n);
+extern size_t file_read(ang_file *f, char *buf, size_t n);
 
 /*
  * Write the first `n` bytes following the pointer `buf` to the file represented
